bugfix: the meaning of '<+>' was misunderstood in TemplateParser.
'<+>' in hxt is not to be understood as a logical 'or'. Instead the hxt
arrow need to be understood as filters. a1 <+> a2 is the filter that
passes through filter a1 or filter a2. For example  a1 <+> const A
Nothing will match always! So the result of runX (a1 <+> a2) will not be
either the result of a1 'or' a2, but both of them collected in a list.