feature: Got rid of the bound checks.
The bound checks would never have catched an over/underflow anyway,
since we are using Int32/Word32 types and they would silently overflow
before the bound checking occurs.
This commit doesn't compile, because the + operation is not defined on
Int primivites. But these will be replaced anyway in the next commit.